refactor: components of filter
