+++ lib/components/filter/filterPage.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:postin_app/components/filter/components/searchRange.dart';
+import 'package:postin_app/components/filter/components/priceRangeSliderWidget.dart';
+import 'package:postin_app/components/filter/components/popularityWidget.dart';
+import 'package:postin_app/components/filter/components/searchRangeWidget.dart';
+import 'package:postin_app/components/filter/components/rangeDatesWidget.dart';
 
 class EventsFilterPage extends StatefulWidget {
   @override
@@ -105,204 +108,4 @@ class _EventsFilterPageState extends State<EventsFilterPage> {
       ),
     );
   }
-}
-
-Widget searchRange(
-  BuildContext context,
-  int _searchRange,
-  Function(int) onRangeChanged,
-) {
-  return Column(
-    crossAxisAlignment: CrossAxisAlignment.start,
-    children: [
-      Text(
-        'Search Range',
-        style: TextStyle(
-          fontSize: 20,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-      SizedBox(height: 10),
-      GestureDetector(
-        onTap: () async {
-          // Navigate to another screen
-          final result = await Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => SearchRange(
-                initialRangeValue: _searchRange,
-              ),
-            ),
-          );
-
-          if (result != null) {
-            onRangeChanged(int.parse(result.toString()));
-          }
-        },
-        child: Row(
-          children: [
-            Icon(Icons.location_on_outlined, size: 30),
-            SizedBox(width: 10),
-            Text(
-              'View search range: $_searchRange km',
-              style: TextStyle(
-                color: Theme.of(context).primaryColor,
-                fontSize: 20,
-              ),
-            ),
-          ],
-        ),
-      ),
-      SizedBox(height: 10),
-    ],
-  );
-}
-
-Widget rangeDates(
-  BuildContext context,
-  RangeValues _rangeDates,
-  Function(RangeValues) onDateChanged,
-) {
-  // Convertimos las fechas en formato legible
-  String _formattedDate(DateTime date) {
-    return '${date.day}/${date.month}/${date.year}';
-  }
-
-  return Column(
-    crossAxisAlignment: CrossAxisAlignment.start,
-    children: [
-      Text(
-        'Date Range',
-        style: TextStyle(
-          fontSize: 20,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-      SizedBox(height: 5),
-      GestureDetector(
-        onTap: () async {
-          final DateTimeRange? picked = await showDateRangePicker(
-            context: context,
-            firstDate: DateTime.now(),
-            lastDate: DateTime.now().add(Duration(days: 365 * 100)),
-            initialDateRange: DateTimeRange(
-              start: DateTime.now(),
-              end: DateTime.now().add(Duration(days: 31)),
-            ),
-          );
-          if (picked != null) {
-            onDateChanged(RangeValues(
-              picked.start.millisecondsSinceEpoch.toDouble(),
-              picked.end.millisecondsSinceEpoch.toDouble(),
-            ));
-          }
-        },
-        child: Row(
-          children: [
-            Icon(Icons.date_range, size: 30),
-            Column(
-              children: [
-                SizedBox(width: 10),
-                Text(
-                  '  Select date range',
-                  style: TextStyle(
-                    color: Theme.of(context).primaryColor,
-                    fontSize: 20,
-                  ),
-                ),
-                Text(
-                  '    ${_formattedDate(DateTime.fromMillisecondsSinceEpoch(_rangeDates.start.toInt()))} - ${_formattedDate(DateTime.fromMillisecondsSinceEpoch(_rangeDates.end.toInt()))}',
-                  style: TextStyle(
-                    fontSize: 16,
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    ],
-  );
-}
-
-Widget priceRangeSlider(BuildContext context, RangeValues _priceRange,
-    Function(RangeValues) onChanged) {
-  String rangPrice;
-  int max = 10000;
-
-  if (_priceRange.start == 0 && _priceRange.end == 0) {
-    rangPrice = 'Free';
-  } else if (_priceRange.start == 0 && _priceRange.end == max) {
-    rangPrice = '0 - No limit';
-  } else {
-    rangPrice = '\$${_priceRange.start.round()} - \$${_priceRange.end.round()}';
-  }
-
-  return Column(
-    crossAxisAlignment: CrossAxisAlignment.start,
-    children: [
-      Text(
-        'Price Range:',
-        style: TextStyle(
-          fontSize: 20,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-      Text(
-        rangPrice,
-        style: TextStyle(
-          fontSize: 18,
-          fontWeight: FontWeight.normal,
-        ),
-      ),
-      SizedBox(height: 10),
-      SliderTheme(
-        data: SliderThemeData(
-          trackHeight: 2.0, // Ancho de la barra del slider
-          activeTrackColor: Theme.of(context)
-              .primaryColor, // Color principal de la aplicación
-        ),
-        child: RangeSlider(
-          values: _priceRange,
-          min: 0,
-          max: max * 1,
-          divisions: max,
-          onChanged: onChanged,
-          labels: RangeLabels(
-            '\$${_priceRange.start.round()}',
-            '\$${_priceRange.end.round()}',
-          ),
-          activeColor: Theme.of(context)
-              .primaryColor, // Color principal de la aplicación
-        ),
-      ),
-    ],
-  );
-}
-
-Widget popularity(bool _isPopularSelected, Function(bool) onChanged) {
-  return Column(
-    crossAxisAlignment:
-        CrossAxisAlignment.start, // Alinea el contenido a la izquierda
-    children: [
-      Row(
-        mainAxisAlignment: MainAxisAlignment
-            .spaceBetween, // Alinea los elementos a los extremos
-        children: [
-          Text(
-            'Popularity',
-            style: TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-          Switch(
-            // Utiliza un Switch en lugar de un SwitchListTile para tener más control
-            value: _isPopularSelected,
-            onChanged: onChanged,
-          ),
-        ],
-      ),
-    ],
-  );
 }
